fibonacci sequence up to the 50th term

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,9 +7,9 @@
 
 int main(void)
 {
-int a = 1;
-int b = 2;
-unsigned long c; /* Next number */
+long a = 1;
+long b = 2;
+long c; /* Next number */
 int i; /* Loop variable */
 
 printf("%d, ", a);
@@ -19,7 +19,7 @@ for (i = 2; i < 50; i++)
 {
 c = a + b;
 
-printf("%lu, ", c);
+printf("%ld, ", c);
 
 a = b;
 b = c;
